requirements.txt
streamlit==1.22.0
python-dotenv==1.2.1